+++ test/LokiDbImplTest.js
@@ -54,8 +54,8 @@ var dataModelsMap = dataModels;
 QUnit.module("LokiDbImpl", {});
 QUnit.test("new LokiDbImpl", function LokiDbImplTest(sr) {
     var persister;
-    var dbInst = new LokiDbImpl("lokijs-collections-test", { readAllow: true, writeAllow: true }, { compressLocalStores: false }, "collection_meta_data", ModelDefinitionsSet.fromCollectionModels(dataModelsMap, dataTypes), function createPersister(dbInst) {
-        persister = new DummyDataPersister(function () { return dbInst.getCollections(); }, LokiDbImpl.stripMetaData, null);
+    var dbInst = new LokiDbImpl("lokijs-collections-test", { readAllow: true, writeAllow: true }, { compressLocalStores: false }, "for-in-if", "collection_meta_data", ModelDefinitionsSet.fromCollectionModels(dataModelsMap, dataTypes), function createPersister(dbInst) {
+        persister = new DummyDataPersister(function () { return dbInst.getCollections(); }, LokiDbImpl.cloneWithoutMetaData_for_in_if, null);
         return persister;
     });
     dbInst.initializeDb({});