+++ test/LokiDbImplTest.ts
@@ -99,10 +99,10 @@ QUnit.module("LokiDbImpl", {
 
 QUnit.test("new LokiDbImpl", function LokiDbImplTest(sr) {
     var persister: DummyDataPersister;
-    var dbInst = new LokiDbImpl("lokijs-collections-test", { readAllow: true, writeAllow: true }, { compressLocalStores: false },
+    var dbInst = new LokiDbImpl("lokijs-collections-test", { readAllow: true, writeAllow: true }, { compressLocalStores: false }, "for-in-if",
         "collection_meta_data", ModelDefinitionsSet.fromCollectionModels(dataModelsMap, dataTypes),
         function createPersister(dbInst: InMemDb) {
-            persister = new DummyDataPersister(() => dbInst.getCollections(), LokiDbImpl.stripMetaData, null);
+            persister = new DummyDataPersister(() => dbInst.getCollections(), LokiDbImpl.cloneWithoutMetaData_for_in_if, null);
             return persister;
         }
     );